ic_ctwo/requirements.txt
pika==1.2.0
Flask==2.1.2
mysql-connector-python==8.0.23